feat: Add consolidated _reference/ folder with setup skill
- Create _reference/ folder with README, .gitignore, and .setup.json.example
- Add /setup-references skill for configuring reference code sources
- Track user choices (clone vs symlink) in .setup.json
- Fix mypy type errors (numpy floating, Optional[float], numpy.bool)
- Add mypy overrides to skip pytest module imports
- Disable overly strict mypy checks (var-annotated, warn_return_any)
- Update CLAUDE.md with _reference/ documentation

Supports both git clones and symlinks for reference code like
SlicerSource, SlicerSegmentEditorExtraEffects, and QuantitativeReporting.

+++ SegmentEditorAdaptiveBrush/SegmentEditorAdaptiveBrushLib/IntensityAnalyzer.py
@@ -98,7 +98,7 @@ class IntensityAnalyzer:
         except IndexError:
             logging.warning(f"Seed point {seed_point} out of bounds")
             # Return conservative thresholds based on full image
-            return self._simple_statistics(image.flatten(), np.mean(image), edge_sensitivity)
+            return self._simple_statistics(image.flatten(), float(np.mean(image)), edge_sensitivity)
 
         # Use GMM if available and requested
         if self.use_gmm:
